src/requirements.txt
py-xmltv==1.0.8
lxml==4.9.3
requests
html5lib
beautifulsoup4